requirements.txt
python-gnupg
pydantic
pydantic_settings
requests
sqlalchemy
